requirements.txt
asgiref==3.5.2
attrs==20.3.0
cachelib==0.9.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
cssbeautifier==1.14.6
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djlint==1.12.1
EditorConfig==0.12.3
filelock==3.8.0
fonttools==4.37.4
fpdf2==2.5.7
greenlet==1.1.3
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
importlib-metadata==4.12.0
inflect==6.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
jellyfish==0.9.0
Jinja2==3.1.2
jsbeautifier==1.14.6
lib50==3.0.4
MarkupSafe==2.1.1
packaging==21.3
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.5
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.1.3
pytz==2022.6
PyYAML==5.4.1
regex==2022.8.17
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlparse==0.4.2
svg.path==6.2
termcolor==1.1.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.6
Werkzeug==2.2.2
whitenoise==6.2.0
zipp==3.8.1
